Remove unnecessary casts

+++ modules/javafx.graphics/src/main/java/com/sun/javafx/font/PrismFontFactory.java
@@ -1927,7 +1927,7 @@ public abstract class PrismFontFactory implements FontFactory {
             } else if (isEmbedded) {
                 try {
                     int screenDPI = Screen.getMainScreen().getResolutionY();
-                    systemFontSize = ((float) screenDPI) / 6f; // 12 points
+                    systemFontSize = screenDPI / 6f; // 12 points
                 } catch (NullPointerException npe) {
                     // if no screen is defined
                     systemFontSize = 13f; // same as desktop Linux
