+++ modules/javafx.graphics/src/main/java/com/sun/javafx/util/Utils.java
@@ -387,7 +387,7 @@ public class Utils {
 
         brightness = cmax;
         if (cmax != 0)
-            saturation = (double) (cmax - cmin) / cmax;
+            saturation = (cmax - cmin) / cmax;
         else
             saturation = 0;
